Warning--

+++ src/kcompletionbox.cpp
@@ -372,7 +372,7 @@ QRect KCompletionBox::calculateGeometry() const
 
     int x = 0, y = 0;
     int ih = visualRect.height();
-    int h = qMin(15 * ih, (int) count() * ih) + 2 * frameWidth();
+    int h = qMin(15 * ih, count() * ih) + 2 * frameWidth();
 
     int w = (d->m_parent) ? d->m_parent->width() : QListWidget::minimumSizeHint().width();
     w = qMax(QListWidget::minimumSizeHint().width(), w);
